// src/app/models/checkout/CheckoutReturnState.ts
export interface CheckoutReturnState {
   vehicleFrontView: string;
   vehicleBackView: string;
   vehicleRightSide: string;
   vehicleLeftSide: string;
   vehiclePlateNumber: string;
   iban: string;
   cardBrand: string;
   cardNumber: string;
   cardHolder: string;
   cardExpiryDate: string;
   cardCVV: string;
}